Add form and display functions
The page now starts with a form for users to submit their names,
when submitted, the form disappears and the board appears.
As the game finishes, a ending message & reset button appears. The
reset button still doesn't work, which is what I will be working on
in the next commit