Fix checkWin bug with tie
The checkWin function uses a count that increases every time it
is called to count the turns. Since it was being called multiple times
the cound wasn't properly counting the turns. By assigning it to a
'winner' constant, it can be only called once per turn, making
the count work effectively and the tie functionality to work properly